Save progress for cluster.reset()

+++ lib/cluster.py
@@ -15,6 +15,7 @@ from lib.sgaccel import SgAccel
 from lib.server import Server
 from lib.admin import Admin
 from lib import settings
+import lib.constants
 from provision.ansible_runner import run_ansible_playbook
 
 import logging
@@ -71,7 +72,7 @@ class Cluster:
         if len(self.sync_gateways) == 0:
             raise Exception("Functional tests require at least 1 index reader")
 
-    def reset(self, config):
+    def reset(self, config, run_opts):
 
         self.validate_cluster()
         
@@ -85,25 +86,27 @@ class Cluster:
         status = run_ansible_playbook("stop-sg-accel.yml", stop_on_fail=False)
         assert(status == 0)
 
-        # Deleting sync_gateway artifacts
-        log.info(">>> Deleting sync_gateway artifacts")
-        status = run_ansible_playbook("delete-sync-gateway-artifacts.yml", stop_on_fail=False)
-        assert(status == 0)
+        mode = lib.constants.channel_cache
+        conf_path = os.path.abspath("conf/" + config)
 
-        # Deleting sg_accel artifacts
-        log.info(">>> Deleting sg_accel artifacts")
-        status = run_ansible_playbook("delete-sg-accel-artifacts.yml", stop_on_fail=False)
-        assert(status == 0)
+        if run_opts.reset_data:
+            # Deleting sync_gateway artifacts
+            log.info(">>> Deleting sync_gateway artifacts")
+            status = run_ansible_playbook("delete-sync-gateway-artifacts.yml", stop_on_fail=False)
+            assert(status == 0)
 
-        # Delete buckets
-        log.info(">>> Deleting buckets on: {}".format(self.servers[0].ip))
-        self.servers[0].delete_buckets()
+            # Deleting sg_accel artifacts
+            log.info(">>> Deleting sg_accel artifacts")
+            status = run_ansible_playbook("delete-sg-accel-artifacts.yml", stop_on_fail=False)
+            assert(status == 0)
+
+            # Delete buckets
+            log.info(">>> Deleting buckets on: {}".format(self.servers[0].ip))
+            self.servers[0].delete_buckets()
 
         # Parse config and grab bucket names
-        conf_path = os.path.abspath("conf/" + config)
         bucket_names_from_config = []
 
-        mode = "channel_cache"
         with open(conf_path, "r") as config:
             data = config.read()
 
@@ -118,7 +121,7 @@ class Cluster:
 
             # Add CBGT buckets
             if "cluster_config" in conf_obj.keys():
-                mode = "distributed_index"
+                mode = lib.constants.distributed_index
                 bucket_names_from_config.append(conf_obj["cluster_config"]["bucket"])
 
             dbs = conf_obj["databases"]
@@ -132,9 +135,18 @@ class Cluster:
         # Buckets may be shared for different functionality
         bucket_name_set = list(set(bucket_names_from_config))
 
-        log.info(">>> Creating buckets on: {}".format(self.servers[0].ip))
-        log.info(">>> Creating buckets {}".format(bucket_name_set))
-        self.servers[0].create_buckets(bucket_name_set)
+        # Assert that conf author is alerted if they do not follow conventions for bucket naming
+        if run_opts.mode == lib.constants.distributed_index:
+            assert len(bucket_name_set) == 1 and "data_bucket" in bucket_name_set
+        elif run_opts.mode == lib.constants.channel_cache:
+            assert len(bucket_name_set) == 2 and "data_bucket" in bucket_name_set and "index_bucket" in bucket_name_set
+        else:
+            raise ValueError("Unsupported mode")
+
+        if run_opts.reset_data:
+            log.info(">>> Creating buckets on: {}".format(self.servers[0].ip))
+            log.info(">>> Creating buckets {}".format(bucket_name_set))
+            self.servers[0].create_buckets(bucket_name_set)
 
         log.info(">>> Starting sync_gateway with configuration: {}".format(conf_path))
 
@@ -148,7 +160,7 @@ class Cluster:
 
         # HACK - only enable sg_accel for distributed index tests
         # revise this with https://github.com/couchbaselabs/sync-gateway-testcluster/issues/222
-        if mode == "distributed_index":
+        if mode == lib.constants.distributed_index:
             # Start sg-accel
             status = run_ansible_playbook(
                 "start-sg-accel.yml",
@@ -158,7 +170,7 @@ class Cluster:
             assert(status == 0)
 
         # Validate CBGT
-        if mode == "distributed_index":
+        if mode == lib.constants.distributed_index:
             if not self.validate_cbgt_pindex_distribution_retry():
                 self.save_cbgt_diagnostics()
                 raise Exception("Failed to validate CBGT Pindex distribution")
@@ -273,7 +285,7 @@ class Cluster:
                 log.error("sync_gateway down: {}".format(e))
                 errors.append((sg, e))
 
-        if mode == "distributed_index":
+        if mode == lib.constants.distributed_index:
             for sa in self.sg_accels:
                 try:
                     info = sa.info()
